Make draw routines process bytes; let compiler optimize memory ops
The original, hand-optimized operations on chunks of 32-bit words in
sprite routines used to segfault with GCC 7.5 at -O3.

The problem was that the 32-bit words were likely to be misaligned,
depending on the position of the sprite in the framebuffer. At -O3, GCC
7.5 would do SSE loads from the misaligned words in memory, which would
crash. GCC 8 and later do not appear to exhibit the issue.

To work around the issue, I removed the hand optimizations from the
sprite draw routines; the only data size they iterate on is now bytes.
This appears to fix the crash with GCC 7.5 at -O3.

Removing these hand optimizations is fine performance-wise because
modern compilers are smart enough to generate the appropriate
optimizations to work on bigger data sizes than bytes when possible.